Responsivo a telefono + owner modal + boton volver

+++ src/main/resources/static/js/logout.js
@@ -1,7 +1,16 @@
 (function() {
 	const btn = document.querySelector("#btn-cerrar-sesion");
-	btn.addEventListener("click", function() {
-		sessionStorage.clear();
-		document.location = `${BASE_URL}app/cerrar`;
-	})
+	if(btn){
+		btn.addEventListener("click", function() {
+			sessionStorage.clear();
+			document.location = `${BASE_URL}app/cerrar`;
+		})
+	}
+	const btn_phone = document.querySelector("#btn-cerrar-sesion-phone");
+	if(btn_phone){
+		btn_phone.addEventListener("click", function() {
+			sessionStorage.clear();
+			document.location = `${BASE_URL}app/cerrar`;
+		})
+	}
 })();
